move UserReactions into components/molecules
