+++ src/components/molecules/UserReactions/index.ts
@@ -1,0 +1,1 @@
+export { UserReactions } from "./UserReactions";
